html/semantics/document-metadata/the-style-element/style-attributes.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Miscellaneous tests for style element attributes</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dev.w3.org/csswg/cssom/#extensions-to-the-document-interface">
    <style id="test2">
      #test1 {
        display: inline;
        background: white;
      }
    </style>
    <style id="test3" media="all" type="text/css" title="alternate1">
      #test1 {
        background: blue;
      }
    </style>
    <style title="alternate2">
      #test1 {
        background: yellow;
      }
    </style>
  </head>
  <body>
    <div id="log"></div>
    <div id="test1">x</div>
    <script>
      test(function() {
        assert_equals(document.styleSheetSets[0], "alternate1", "Alternate stylesheet was not reconized as such");
      }, "The title attribute defines an alternative style sheet");
      test(function() {
        var test2 = document.getElementById("test2");
        var test3 = document.getElementById("test3");
        assert_equals(test2.media, "", "Omitted media value was not IDL reflected as empty string");
        assert_equals(test2.type, "", "Omitted type value was not IDL reflected as empty string");
        assert_equals(test3.media, "all", "Assigned (and valid) media value was not IDL reflected");
        assert_equals(test3.type, "text/css", "Assigned (and valid) type value was not IDL reflected");
      }, "The media and type IDL attributes reflect their respective content attributes");
    </script>
  </body>
</html>
